Restructure UsageStore and split SmartPanels.swift, no behavior change
The onboarding and AI-insight features tacked more responsibilities onto
UsageStore (config/onboarding, refresh, sessions, notifications, AI insight,
debug logging all in one 685-line file) and dumped their new views into
SmartPanels.swift alongside unrelated History/Event/Config panels.

Split UsageStore into per-concern extension files (Onboarding, Refresh,
Accounts, Sessions, Notifications, Insight, Debug, Agents) and broke
SmartPanels.swift into one file per view (HistoryPanel, EventPanel,
SettingsPanel, ConfigEditor, AIInstructionsEditor), matching the
one-type-per-file convention already used elsewhere (AccountCard,
OnboardingView). EmptyPanel moved alongside the other shared UI
primitives in Components.swift. Pure reorganization, verified with a
clean swift build after each step.
