+++ Sources/Toki/Store/UsageStore+Accounts.swift
@@ -1,0 +1,67 @@
+import Foundation
+
+extension UsageStore {
+    func adjustUsage(accountID: String, delta: Double) {
+        guard let account = config?.accounts.first(where: { $0.id == accountID }), account.provider.isConsumerTracked else {
+            return
+        }
+        let current = usageState.accounts[accountID]?.used ?? account.used ?? usageFromRemaining(account)
+        let next = max(current + delta, 0)
+        usageState.accounts[accountID] = AccountUsageState(
+            used: next,
+            lastResetAt: usageState.accounts[accountID]?.lastResetAt ?? resetAnchorDate(for: account)
+        )
+        StateLoader.save(usageState)
+        refresh()
+    }
+
+    func resetUsage(accountID: String) {
+        guard let account = config?.accounts.first(where: { $0.id == accountID }), account.provider.isConsumerTracked else {
+            return
+        }
+        usageState.accounts[accountID] = AccountUsageState(used: 0, lastResetAt: Date())
+        StateLoader.save(usageState)
+        refresh()
+    }
+
+    func renameAccount(snapshot: AccountSnapshot, alias: String) {
+        let trimmed = alias.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !trimmed.isEmpty, var config else { return }
+
+        if snapshot.provider.isClaudeAccount,
+           let email = emailAddress(in: snapshot) {
+            var labels = config.accountLabels ?? []
+            if let index = labels.firstIndex(where: { $0.email.caseInsensitiveCompare(email) == .orderedSame }) {
+                labels[index].nickname = trimmed
+            } else {
+                labels.append(AccountLabelConfig(
+                    email: email,
+                    organizationUuid: nil,
+                    organizationName: organizationName(in: snapshot),
+                    nickname: trimmed,
+                    emoji: nil,
+                    color: nil
+                ))
+            }
+            config.accountLabels = labels
+        } else if let index = config.accounts.firstIndex(where: { $0.id == snapshot.id }) {
+            config.accounts[index].name = trimmed
+        } else {
+            return
+        }
+
+        do {
+            try ConfigLoader.save(config)
+            self.config = config
+            snapshots = snapshots.map { current in
+                guard current.id == snapshot.id else { return current }
+                var updated = current
+                updated.name = trimmed
+                return updated
+            }
+        } catch {
+            DiagnosticLogger.shared.record(.error, component: "config", code: "alias_save_failed", detail: diagnosticErrorDetail(error))
+            configError = "Could not save alias: \(error.localizedDescription)"
+        }
+    }
+}
