+++ Sources/Toki/Store/UsageStore.swift
@@ -1,7 +1,4 @@
-import AppKit
 import Foundation
-import SwiftUI
-@preconcurrency import UserNotifications
 
 @MainActor
 final class UsageStore: ObservableObject {
@@ -30,15 +27,15 @@ final class UsageStore: ObservableObject {
     @Published var isScanningProviders = false
     @Published private(set) var needsOnboarding = false
 
-    private var config: AppConfig?
-    private var usageState = UsageState()
-    private var timer: Timer?
-    private var isRefreshing = false
-    private var isScanningAgents = false
-    private var eventGeneration = 0
-    private var insightGeneration = 0
-    private var notificationAuthorization: Bool?
-    private var agentTimer: Timer?
+    var config: AppConfig?
+    var usageState = UsageState()
+    var timer: Timer?
+    var isRefreshing = false
+    var isScanningAgents = false
+    var eventGeneration = 0
+    var insightGeneration = 0
+    var notificationAuthorization: Bool?
+    var agentTimer: Timer?
 
     init() {
         reloadConfig()
@@ -48,638 +45,25 @@ final class UsageStore: ObservableObject {
         }
     }
 
-    func refreshActiveAgents() {
-        // Serialize scans: a slow scan must not overlap the next tick, or two detached
-        // tasks would race the scanner's PID cache. isScanningAgents is MainActor-isolated.
-        guard !isScanningAgents else { return }
-        isScanningAgents = true
-        Task {
-            activeAgents = await ActiveAgentScanner.scan()
-            isScanningAgents = false
-        }
-    }
-
     var refreshInterval: TimeInterval {
         TimeInterval(max(config?.refreshMinutes ?? 5, 1) * 60)
     }
 
-    // True when there's nothing to lose by showing the connect wizard: no config.json yet,
-    // or one that decodes fine but simply has no accounts. A config.json that exists and
-    // decodes but is invalid for any other reason (e.g. a copilot entry) - or doesn't decode
-    // at all - is a real error; connect() would still fail its own validate() call there, so
-    // showing Connect buttons would be a dead end. Those cases fall through to the normal
-    // error banner instead (config stays nil, needsOnboarding is false).
-    //
-    // Computed once per reloadConfig() and cached in the stored property above rather than
-    // recomputed as a view-body computed property - it does synchronous disk I/O + JSON
-    // decoding, which shouldn't run on every SwiftUI re-render.
-    private func computeNeedsOnboarding() -> Bool {
-        guard FileManager.default.fileExists(atPath: ConfigLoader.path) else { return true }
-        return ConfigLoader.loadRawIfParsable()?.accounts.isEmpty ?? false
-    }
-
-    func reloadConfig() {
-        do {
-            config = try ConfigLoader.load()
-            needsOnboarding = false
-            usageState = StateLoader.load()
-            syncPublishedState()
-            applyScheduledResets()
-            configError = nil
-            snapshots = config?.accounts.map(AccountSnapshot.loading) ?? []
-            updateDerivedState(for: snapshots)
-            scheduleRefresh()
-            refresh(keepsExistingSnapshots: true, minimumRefreshInterval: 60)
-        } catch {
-            DiagnosticLogger.shared.record(.error, component: "config", code: "load_failed", detail: diagnosticErrorDetail(error))
-            config = nil
-            needsOnboarding = computeNeedsOnboarding()
-            snapshots = []
-            updateDerivedState(for: snapshots)
-            // The onboarding state (missing/empty config) is expected, not an error - don't
-            // show a scary "Missing config" message for it. Genuinely broken configs still do.
-            if needsOnboarding {
-                configError = nil
-                scanForOnboarding()
-            } else {
-                configError = error.localizedDescription
-            }
-        }
-    }
-
-    // Re-probes for newly installed/authenticated CLIs while the onboarding screen is
-    // showing, so e.g. signing into Codex after Toki launched doesn't require a restart.
-    // Called each time the popover opens; a no-op once a config exists.
-    func rescanProvidersIfNeeded() {
-        guard needsOnboarding else { return }
-        scanForOnboarding()
-    }
-
-    // Probes for installed/authenticated CLIs (Claude Code, Codex, OpenCode) so the
-    // onboarding screen can offer them as one-click connects.
-    private func scanForOnboarding() {
-        guard !isScanningProviders else { return }
-        isScanningProviders = true
-        Task {
-            detectedProviders = await ProviderDetection.scan()
-            isScanningProviders = false
-        }
-    }
-
-    // Appends detected accounts to config.json (creating it if this is a fresh install)
-    // and reloads. Accounts already present for the same provider+id are skipped.
-    //
-    // config is nil for ANY load failure, including the exact onboarding case of a file
-    // that decodes fine but has no accounts yet - that's safe to build on top of via
-    // loadRawIfParsable(). A file that exists but doesn't decode at all must not be
-    // silently replaced with a blank one (data loss), so that case alone is refused.
-    func connect(_ accounts: [AccountConfig]) {
-        var next: AppConfig
-        if let config {
-            next = config
-        } else if let parsed = ConfigLoader.loadRawIfParsable() {
-            next = parsed
-        } else if !FileManager.default.fileExists(atPath: ConfigLoader.path) {
-            next = AppConfig(refreshMinutes: nil, accountLabels: nil, accounts: [], aiInstructions: nil)
-        } else {
-            configError = "\(ConfigLoader.path) exists but couldn't be parsed - fix or replace it with the config editor before connecting."
-            return
-        }
-        var existingKeys = Set(next.accounts.map { "\($0.provider.rawValue):\($0.id)" })
-        for account in accounts {
-            let key = "\(account.provider.rawValue):\(account.id)"
-            guard !existingKeys.contains(key) else { continue }
-            next.accounts.append(account)
-            existingKeys.insert(key)
-        }
-        do {
-            try ConfigLoader.validate(next)
-            try ConfigLoader.save(next)
-            reloadConfig()
-        } catch {
-            DiagnosticLogger.shared.record(.error, component: "config", code: "connect_failed", detail: diagnosticErrorDetail(error))
-            configError = "Could not connect account: \(error.localizedDescription)"
-        }
-    }
-
-    func refresh(keepsExistingSnapshots: Bool = true, minimumRefreshInterval: TimeInterval? = nil) {
-        guard let config, !isRefreshing else {
-            if isRefreshing { logDebug("Refresh skipped - already in progress") }
-            return
-        }
-        isRefreshing = true
-        logDebug("Refresh started")
-        if !keepsExistingSnapshots || snapshots.isEmpty {
-            snapshots = config.accounts.map(AccountSnapshot.loading)
-        }
-        let currentState = usageState
-        let previousSnapshots = snapshots
-
-        Task {
-            defer { isRefreshing = false }
-            let response = await UsageFetcher.fetch(
-                config: config,
-                state: currentState,
-                previousSnapshots: previousSnapshots,
-                minimumRefreshInterval: minimumRefreshInterval
-            )
-            for key in response.apiCallKeys {
-                usageState.apiLastCalledAt[key] = response.fetchedAt
-            }
-            if !response.apiCallKeys.isEmpty {
-                StateLoader.save(usageState)
-            }
-            let sorted = sortedByAvailability(response.snapshots)
-            let errorCount = sorted.filter(\.isError).count
-            logDebug("Refresh complete: \(sorted.count) accounts (\(errorCount) errors)")
-            for snapshot in sorted where snapshot.isError {
-                logDebug("  \(snapshot.provider.displayName): unavailable")
-            }
-            recordHistory(for: sorted, at: response.fetchedAt)
-            evaluateEventsAndNotifications(for: sorted, previous: previousSnapshots, at: response.fetchedAt)
-            pruneState(now: response.fetchedAt)
-            StateLoader.save(usageState)
-            withAnimation(.spring(response: 0.32, dampingFraction: 0.86)) {
-                snapshots = sorted
-            }
-            lastUpdated = Date()
-            updateDerivedState(for: sorted)
-        }
-    }
-
-    func adjustUsage(accountID: String, delta: Double) {
-        guard let account = config?.accounts.first(where: { $0.id == accountID }), account.provider.isConsumerTracked else {
-            return
-        }
-        let current = usageState.accounts[accountID]?.used ?? account.used ?? usageFromRemaining(account)
-        let next = max(current + delta, 0)
-        usageState.accounts[accountID] = AccountUsageState(
-            used: next,
-            lastResetAt: usageState.accounts[accountID]?.lastResetAt ?? resetAnchorDate(for: account)
-        )
-        StateLoader.save(usageState)
-        refresh()
-    }
-
-    func resetUsage(accountID: String) {
-        guard let account = config?.accounts.first(where: { $0.id == accountID }), account.provider.isConsumerTracked else {
-            return
-        }
-        usageState.accounts[accountID] = AccountUsageState(used: 0, lastResetAt: Date())
-        StateLoader.save(usageState)
-        refresh()
-    }
-
-    func renameAccount(snapshot: AccountSnapshot, alias: String) {
-        let trimmed = alias.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty, var config else { return }
-
-        if snapshot.provider.isClaudeAccount,
-           let email = emailAddress(in: snapshot) {
-            var labels = config.accountLabels ?? []
-            if let index = labels.firstIndex(where: { $0.email.caseInsensitiveCompare(email) == .orderedSame }) {
-                labels[index].nickname = trimmed
-            } else {
-                labels.append(AccountLabelConfig(
-                    email: email,
-                    organizationUuid: nil,
-                    organizationName: organizationName(in: snapshot),
-                    nickname: trimmed,
-                    emoji: nil,
-                    color: nil
-                ))
-            }
-            config.accountLabels = labels
-        } else if let index = config.accounts.firstIndex(where: { $0.id == snapshot.id }) {
-            config.accounts[index].name = trimmed
-        } else {
-            return
-        }
-
-        do {
-            try ConfigLoader.save(config)
-            self.config = config
-            snapshots = snapshots.map { current in
-                guard current.id == snapshot.id else { return current }
-                var updated = current
-                updated.name = trimmed
-                return updated
-            }
-        } catch {
-            DiagnosticLogger.shared.record(.error, component: "config", code: "alias_save_failed", detail: diagnosticErrorDetail(error))
-            configError = "Could not save alias: \(error.localizedDescription)"
-        }
-    }
-
-    // Custom on-device AI prompt, or nil when the built-in default instructions apply.
-    var aiInstructions: String? {
-        config?.aiInstructions
-    }
-
-    // Whether the on-device AI insight can run at all (macOS 26 + Apple Intelligence),
-    // so Settings can hide the instructions editor on systems that never use it.
-    var isAIInsightAvailable: Bool {
-        #if canImport(FoundationModels)
-        if #available(macOS 26, *) { return InsightGenerator.isAvailable }
-        #endif
-        return false
-    }
-
-    // Persists custom AI instructions (empty clears back to the default) and regenerates
-    // the insight immediately so the user sees the effect of their prompt. Returns the error
-    // on failure so the editor can surface it inline instead of flashing a false "Saved".
-    func updateAIInstructions(_ text: String) -> String? {
-        guard var config else {
-            DiagnosticLogger.shared.record(.error, component: "config", code: "ai_instructions_no_config", detail: "no config loaded")
-            return "No config loaded to save into."
-        }
-        let trimmed = text.trimmingCharacters(in: .whitespacesAndNewlines)
-        config.aiInstructions = trimmed.isEmpty ? nil : trimmed
-        do {
-            try ConfigLoader.save(config)
-            self.config = config
-            refreshAIInsight(for: snapshots)
-            return nil
-        } catch {
-            DiagnosticLogger.shared.record(.error, component: "config", code: "ai_instructions_save_failed", detail: diagnosticErrorDetail(error))
-            return "Could not save AI instructions: \(error.localizedDescription)"
-        }
-    }
-
-    func updatePreferences(_ next: AppPreferences) {
-        preferences = next
-        usageState.preferences = next
-        StateLoader.save(usageState)
-        updateDerivedState(for: snapshots)
-    }
-
-    func setDND(_ isEnabled: Bool) {
-        var next = preferences
-        next.dndEnabled = isEnabled
-        updatePreferences(next)
-        appendEvent(
-            kind: .notification,
-            title: isEnabled ? "DND enabled" : "DND disabled",
-            detail: isEnabled ? "Notifications will be recorded but not delivered." : "Notifications can be delivered again.",
-            deliveredNotification: false
-        )
-    }
-
-    func startSession() {
-        let ratios = Dictionary(uniqueKeysWithValues: snapshots.compactMap { snapshot in
-            snapshot.remainingRatio.map { (snapshot.id, $0) }
-        })
-        let primaries = Dictionary(uniqueKeysWithValues: snapshots.map { ($0.id, $0.primary) })
-        let next = SessionState(startedAt: Date(), startingRemainingRatios: ratios, startingPrimaries: primaries)
-        session = next
-        usageState.session = next
-        StateLoader.save(usageState)
-        appendEvent(kind: .session, title: "Session started", detail: "Toki is tracking quota burn for this coding session.", deliveredNotification: false)
-    }
-
-    func endSession() {
-        guard let session else { return }
-        let summary = sessionSummary(for: session)
-        self.session = nil
-        usageState.session = nil
-        StateLoader.save(usageState)
-        appendEvent(kind: .session, title: "Session ended", detail: summary, deliveredNotification: false)
-    }
-
-    func clearEvents() {
-        eventGeneration += 1
-        events = []
-        usageState.events = []
-        usageState.eventLastRecordedAt = [:]
-        StateLoader.save(usageState)
-    }
-
-    func switchBestAccount() {
-        let current = recommendation
-        guard let target = current.switchTarget else { return }
-        appendEvent(
-            kind: .switchAccount,
-            title: "Smart switch",
-            detail: "Switching to \(recommendedAccountName(from: current)).",
-            deliveredNotification: false
-        )
-        switchClaudeAccount(target: target, command: current.switchCommand)
-    }
-
-    private func recommendedAccountName(from recommendation: SmartRecommendation) -> String {
-        if let accountID = recommendation.accountID,
-           let snapshot = snapshots.first(where: { $0.id == accountID }) {
-            return snapshot.name
-        }
-        return recommendation.title
-            .replacingOccurrences(of: "Switch to ", with: "")
-            .replacingOccurrences(of: "Use ", with: "")
-            .replacingOccurrences(of: " now", with: "")
-    }
-
-    func switchClaudeAccount(target: String, command: String?) {
-        let currentSnapshots = snapshots
-        Task {
-            let result = await Task.detached {
-                Result {
-                    try ClaudeSwapRunner.switchTo(target: target, command: command)
-                }
-            }.value
-
-            switch result {
-            case .success:
-                appendEvent(kind: .switchAccount, title: "Account switched", detail: "Claude Code switched to \(target).", deliveredNotification: false)
-                reloadConfig()
-            case .failure(let error):
-                DiagnosticLogger.shared.record(.error, component: "account_switch", code: "switch_failed", detail: diagnosticErrorDetail(error))
-                appendEvent(kind: .switchAccount, title: "Switch failed", detail: error.localizedDescription, deliveredNotification: false)
-                snapshots = currentSnapshots.map { snapshot in
-                    guard snapshot.switchTarget == target else { return snapshot }
-                    var failed = snapshot
-                    failed.isError = true
-                    failed.metrics = [MetricLine(label: "Switch", value: error.localizedDescription)] + snapshot.metrics
-                    return failed
-                }
-            }
-        }
-    }
-
-    func sessionBurnLines() -> [MetricLine] {
-        guard let session else { return [] }
-        let byID = Dictionary(uniqueKeysWithValues: snapshots.map { ($0.id, $0) })
-        return session.startingRemainingRatios.compactMap { accountID, startingRatio in
-            guard let snapshot = byID[accountID], let current = snapshot.remainingRatio else { return nil }
-            let burned = max(0, startingRatio - current)
-            return MetricLine(label: snapshot.name, value: "\(percentText(burned)) burned")
-        }
-        .sorted { $0.label < $1.label }
-    }
-
-    private func scheduleRefresh() {
+    func scheduleRefresh() {
         timer?.invalidate()
         timer = Timer.scheduledTimer(withTimeInterval: refreshInterval, repeats: true) { [weak self] _ in
             Task { @MainActor in self?.refresh(keepsExistingSnapshots: true) }
         }
     }
 
-    private func applyScheduledResets() {
-        guard let config else { return }
-        var changed = false
-        for account in config.accounts where account.provider.isConsumerTracked {
-            guard let resetEveryHours = account.resetEveryHours, resetEveryHours > 0 else { continue }
-            let currentState = usageState.accounts[account.id]
-            let anchor = currentState?.lastResetAt ?? resetAnchorDate(for: account) ?? Date()
-            let elapsed = Date().timeIntervalSince(anchor)
-            let window = resetEveryHours * 3600
-            if elapsed >= window {
-                let windowsElapsed = floor(elapsed / window)
-                let nextAnchor = anchor.addingTimeInterval(windowsElapsed * window)
-                usageState.accounts[account.id] = AccountUsageState(used: 0, lastResetAt: nextAnchor)
-                changed = true
-            }
-        }
-        if changed {
-            StateLoader.save(usageState)
-        }
-    }
-
-    func logDebug(_ message: String) {
-        guard debugMode else { return }
-        debugLog.append(DebugLogEntry(timestamp: Date(), message: message))
-        if debugLog.count > 100 {
-            debugLog.removeFirst(debugLog.count - 100)
-        }
-    }
-
-    func toggleDebug() {
-        debugMode.toggle()
-        if debugMode {
-            debugLogHandler = { [weak self] in self?.logDebug($0) }
-            logDebug("Debug mode enabled")
-        } else {
-            debugLogHandler = nil
-        }
-    }
-
-    private func syncPublishedState() {
+    func syncPublishedState() {
         preferences = usageState.preferences
         events = usageState.events.sorted { $0.timestamp > $1.timestamp }
         history = usageState.history.sorted { $0.timestamp > $1.timestamp }
         session = usageState.session
     }
 
-    private func updateDerivedState(for snapshots: [AccountSnapshot]) {
-        recommendation = smartRecommendation(for: snapshots)
-        statusEntries = menuBarEntries(for: snapshots, mode: preferences.menuBarMode)
-        if statusEntries.isEmpty {
-            statusEntries = menuBarPlaceholderEntries()
-        }
-        syncPublishedState()
-        refreshAIInsight(for: snapshots)
-    }
-
-    // Enriches the overview with an on-device LLM summary when Apple Intelligence is
-    // available. The deterministic recommendation is already shown; this only replaces
-    // it once ready, and stays nil (rule-based visible) on older systems or failure.
-    private func refreshAIInsight(for snapshots: [AccountSnapshot]) {
-        // FoundationModels only exists in the macOS 26 SDK; when building against an older
-        // SDK the generator is compiled out entirely and the rule-based recommendation stays.
-        #if canImport(FoundationModels)
-        guard #available(macOS 26, *), InsightGenerator.isAvailable else {
-            aiInsight = nil
-            isGeneratingInsight = false
-            return
-        }
-        let grounding = recommendation
-        let instructions = config?.aiInstructions
-        // Keep the previous summary visible while regenerating; a token guards against a
-        // slower earlier generation overwriting a newer one.
-        insightGeneration += 1
-        let generation = insightGeneration
-        isGeneratingInsight = true
-        Task { @MainActor in
-            let result = await InsightGenerator.generate(snapshots: snapshots, grounding: grounding, instructions: instructions)
-            guard generation == insightGeneration else { return }
-            if let result { aiInsight = result }
-            isGeneratingInsight = false
-        }
-        #else
-        aiInsight = nil
-        isGeneratingInsight = false
-        #endif
-    }
-
-    private func recordHistory(for snapshots: [AccountSnapshot], at date: Date) {
-        let candidates = snapshots.filter { !$0.isError && $0.remainingRatio != nil }
-        for snapshot in candidates {
-            let last = usageState.history
-                .filter { $0.accountID == snapshot.id }
-                .max { $0.timestamp < $1.timestamp }
-            let shouldAppend = last == nil
-                || date.timeIntervalSince(last?.timestamp ?? .distantPast) >= 15 * 60
-                || abs((last?.remainingRatio ?? 0) - (snapshot.remainingRatio ?? 0)) >= 0.02
-            guard shouldAppend else { continue }
-            usageState.history.append(UsageHistoryEntry(
-                timestamp: date,
-                accountID: snapshot.id,
-                accountName: snapshot.name,
-                provider: snapshot.provider,
-                remainingRatio: snapshot.remainingRatio,
-                primary: snapshot.primary
-            ))
-        }
-    }
-
-    private func evaluateEventsAndNotifications(for snapshots: [AccountSnapshot], previous: [AccountSnapshot], at date: Date) {
-        let previousByID = Dictionary(uniqueKeysWithValues: previous.map { ($0.id, $0) })
-        for snapshot in snapshots where !snapshot.isError {
-            guard let ratio = snapshot.remainingRatio else { continue }
-            let previousRatio = previousByID[snapshot.id]?.remainingRatio
-            if ratio <= preferences.lowQuotaThreshold,
-               previousRatio == nil || (previousRatio ?? 1) > preferences.lowQuotaThreshold {
-                notifyOrRecord(
-                    key: "lowQuota:\(snapshot.id)",
-                    kind: .lowQuota,
-                    title: "\(snapshot.name) is low",
-                    detail: "\(snapshot.name) has \(percentText(ratio)) quota remaining.",
-                    at: date
-                )
-            } else if ratio >= preferences.lowQuotaThreshold + 0.20,
-                      let previousRatio,
-                      previousRatio <= preferences.lowQuotaThreshold {
-                notifyOrRecord(
-                    key: "recovered:\(snapshot.id)",
-                    kind: .recovered,
-                    title: "\(snapshot.name) recovered",
-                    detail: "\(snapshot.name) is back to \(percentText(ratio)) remaining.",
-                    at: date
-                )
-            }
-        }
-
-        guard let session else { return }
-        for snapshot in snapshots where !snapshot.isError {
-            guard let current = snapshot.remainingRatio,
-                  let starting = session.startingRemainingRatios[snapshot.id] else { continue }
-            let burned = starting - current
-            if current <= preferences.sessionWarningThreshold || burned >= 0.30 {
-                notifyOrRecord(
-                    key: "session:\(snapshot.id)",
-                    kind: .session,
-                    title: "Session quota warning",
-                    detail: "\(snapshot.name) has \(percentText(current)) left after burning \(percentText(max(0, burned))).",
-                    at: date
-                )
-            }
-        }
-    }
-
-    private func notifyOrRecord(key: String, kind: TokiEventKind, title: String, detail: String, at date: Date) {
-        let cooldown = TimeInterval(max(preferences.notificationCooldownMinutes, 5) * 60)
-        if let last = usageState.eventLastRecordedAt[key],
-           date.timeIntervalSince(last) < cooldown {
-            return
-        }
-
-        let canAttemptDelivery = preferences.notificationsEnabled && !preferences.dndEnabled
-        usageState.eventLastRecordedAt[key] = date
-        guard canAttemptDelivery else {
-            appendEvent(
-                kind: kind,
-                title: title,
-                detail: preferences.dndEnabled ? "DND: \(detail)" : "Notifications disabled: \(detail)",
-                deliveredNotification: false,
-                at: date
-            )
-            return
-        }
-
-        let generation = eventGeneration
-        deliverNotification(title: title, detail: detail) { delivered, failureDetail in
-            guard generation == self.eventGeneration else { return }
-            self.appendEvent(
-                kind: kind,
-                title: title,
-                detail: delivered ? detail : "Not delivered: \(failureDetail ?? detail)",
-                deliveredNotification: delivered,
-                at: date
-            )
-        }
-    }
-
-    private func deliverNotification(
-        title: String,
-        detail: String,
-        completion: @escaping @MainActor (Bool, String?) -> Void
-    ) {
-        resolveNotificationAuthorization { granted, error in
-            if let error {
-                Task { @MainActor in completion(false, error) }
-                return
-            }
-            guard granted else {
-                Task { @MainActor in completion(false, "notification permission denied") }
-                return
-            }
-
-            let content = UNMutableNotificationContent()
-            content.title = title
-            content.body = detail
-            content.sound = .default
-            let request = UNNotificationRequest(identifier: UUID().uuidString, content: content, trigger: nil)
-            UNUserNotificationCenter.current().add(request) { error in
-                Task { @MainActor in completion(error == nil, error?.localizedDescription) }
-            }
-        }
-    }
-
-    private func resolveNotificationAuthorization(completion: @escaping @Sendable (Bool, String?) -> Void) {
-        if notificationAuthorization == true {
-            completion(true, nil)
-            return
-        }
-        // Only cache a granted result. A denied/undetermined state is re-checked so
-        // enabling notifications in System Settings takes effect without a restart.
-        UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .sound]) { [weak self] granted, error in
-            if error == nil && granted {
-                Task { @MainActor in self?.notificationAuthorization = true }
-            }
-            completion(error == nil && granted, error?.localizedDescription)
-        }
-    }
-
-    private func appendEvent(
-        kind: TokiEventKind,
-        title: String,
-        detail: String,
-        deliveredNotification: Bool,
-        at date: Date = Date()
-    ) {
-        usageState.events.append(TokiEvent(
-            timestamp: date,
-            kind: kind,
-            title: title,
-            detail: detail,
-            deliveredNotification: deliveredNotification
-        ))
-        pruneState(now: date)
-        StateLoader.save(usageState)
-        syncPublishedState()
-    }
-
-    private func pruneState(now: Date) {
-        let retention = TimeInterval(max(preferences.historyRetentionDays, 1) * 24 * 60 * 60)
-        usageState.history = Array(usageState.history
-            .filter { now.timeIntervalSince($0.timestamp) <= retention }
-            .suffix(720))
-        usageState.events = Array(usageState.events.suffix(160))
-    }
-
-    private func sessionSummary(for session: SessionState) -> String {
-        let elapsed = formatDuration(seconds: Date().timeIntervalSince(session.startedAt))
-        let lines = sessionBurnLines()
-        guard !lines.isEmpty else { return "Tracked for \(elapsed)." }
-        let top = lines.prefix(2).map { "\($0.label): \($0.value)" }.joined(separator: ", ")
-        return "Tracked for \(elapsed). \(top)."
+    func setNeedsOnboarding(_ value: Bool) {
+        needsOnboarding = value
     }
 }
