--- Sources/Toki/Views/SmartPanels.swift
@@ -1,532 +1,0 @@
-import SwiftUI
-
-struct HistoryPanel: View {
-    @ObservedObject var store: UsageStore
-
-    var body: some View {
-        ScrollView {
-            LazyVStack(alignment: .leading, spacing: 8) {
-                if store.history.isEmpty {
-                    EmptyPanel(systemImage: "chart.line.uptrend.xyaxis", title: "No history yet", detail: "Toki records local quota snapshots after refreshes.")
-                } else {
-                    ForEach(store.history.prefix(80)) { entry in
-                        HStack(alignment: .center, spacing: 8) {
-                            ProviderLogo(provider: entry.provider, size: 16)
-                            VStack(alignment: .leading, spacing: 2) {
-                                Text(entry.accountName)
-                                    .font(.system(size: 12, weight: .semibold))
-                                    .lineLimit(1)
-                                Text(entry.timestamp, format: .dateTime.month(.abbreviated).day().hour().minute())
-                                    .font(.system(size: 10))
-                                    .foregroundStyle(.secondary)
-                            }
-                            Spacer()
-                            VStack(alignment: .trailing, spacing: 2) {
-                                Text(entry.remainingRatio.map(percentText) ?? "--")
-                                    .font(.system(size: 12, weight: .semibold, design: .monospaced))
-                                    .foregroundStyle(entry.remainingRatio.map(historyColor) ?? .secondary)
-                                Text(entry.primary)
-                                    .font(.system(size: 9))
-                                    .foregroundStyle(.tertiary)
-                                    .lineLimit(1)
-                            }
-                        }
-                        .padding(8)
-                        .background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 8, style: .continuous))
-                        .overlay(
-                            RoundedRectangle(cornerRadius: 8, style: .continuous)
-                                .stroke(Color.primary.opacity(0.07), lineWidth: 1)
-                        )
-                    }
-                }
-            }
-            .padding(.trailing, 2)
-        }
-        .frame(maxHeight: accountListHeight())
-    }
-
-    private func historyColor(_ ratio: Double) -> Color {
-        if ratio <= 0.15 { return .red }
-        if ratio <= 0.40 { return .orange }
-        return .primary
-    }
-}
-
-struct EventPanel: View {
-    @ObservedObject var store: UsageStore
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack(spacing: 8) {
-                Button {
-                    store.setDND(!store.preferences.dndEnabled)
-                } label: {
-                    Label(store.preferences.dndEnabled ? "DND On" : "DND Off", systemImage: store.preferences.dndEnabled ? "moon.zzz.fill" : "bell")
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-                .help("Toggle do not disturb")
-                .pointerOnHover()
-
-                Spacer()
-
-                Button {
-                    store.clearEvents()
-                } label: {
-                    Image(systemName: "trash")
-                }
-                .buttonStyle(.plain)
-                .frame(width: 25, height: 25)
-                .background(Color.primary.opacity(0.06), in: RoundedRectangle(cornerRadius: 7, style: .continuous))
-                .help("Clear event history")
-                .accessibilityLabel("Clear event history")
-                .pointerOnHover()
-            }
-
-            ScrollView {
-                LazyVStack(alignment: .leading, spacing: 8) {
-                    if store.events.isEmpty {
-                        EmptyPanel(systemImage: "bell.badge", title: "No events yet", detail: "Low quota, session, switch, and notification events appear here.")
-                    } else {
-                        ForEach(store.events) { event in
-                            HStack(alignment: .top, spacing: 8) {
-                                Image(systemName: icon(for: event))
-                                    .font(.system(size: 13, weight: .semibold))
-                                    .foregroundStyle(color(for: event))
-                                    .frame(width: 18)
-                                VStack(alignment: .leading, spacing: 3) {
-                                    HStack(spacing: 6) {
-                                        Text(event.title)
-                                            .font(.system(size: 12, weight: .semibold))
-                                            .lineLimit(1)
-                                        if event.deliveredNotification {
-                                            Image(systemName: "bell.fill")
-                                                .font(.system(size: 9))
-                                                .foregroundStyle(.blue)
-                                                .accessibilityHidden(true)
-                                        }
-                                    }
-                                    Text(event.detail)
-                                        .font(.system(size: 10))
-                                        .foregroundStyle(.secondary)
-                                        .lineLimit(2)
-                                    Text(event.timestamp, format: .dateTime.month(.abbreviated).day().hour().minute())
-                                        .font(.system(size: 9))
-                                        .foregroundStyle(.tertiary)
-                                }
-                                Spacer()
-                            }
-                            .padding(8)
-                            .background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 8, style: .continuous))
-                            .overlay(
-                                RoundedRectangle(cornerRadius: 8, style: .continuous)
-                                    .stroke(Color.primary.opacity(0.07), lineWidth: 1)
-                            )
-                        }
-                    }
-                }
-                .padding(.trailing, 2)
-            }
-        }
-        .frame(maxHeight: accountListHeight())
-    }
-
-    private func icon(for event: TokiEvent) -> String {
-        switch event.kind {
-        case .lowQuota: return "exclamationmark.triangle.fill"
-        case .recovered: return "checkmark.circle.fill"
-        case .switchAccount: return "arrow.triangle.2.circlepath"
-        case .session: return "timer"
-        case .notification: return event.deliveredNotification ? "bell.fill" : "bell.slash"
-        case .refresh: return "arrow.clockwise"
-        }
-    }
-
-    private func color(for event: TokiEvent) -> Color {
-        switch event.kind {
-        case .lowQuota: return .orange
-        case .recovered: return .green
-        case .switchAccount: return .blue
-        case .session: return .purple
-        case .notification: return event.deliveredNotification ? .blue : .secondary
-        case .refresh: return .secondary
-        }
-    }
-}
-
-// Full-page settings/config view opened from the header gear (no longer a bottom tab).
-struct ConfigPage: View {
-    @ObservedObject var store: UsageStore
-    @ObservedObject var updateChecker: UpdateChecker
-    var onClose: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 10) {
-            HStack(spacing: 8) {
-                Button(action: onClose) {
-                    Image(systemName: "chevron.left")
-                        .font(.system(size: 13, weight: .semibold))
-                        // Fill the whole 25x25 so the entire button surface is the hit
-                        // target, not just the glyph. contentShape makes the padded area tappable.
-                        .frame(width: 25, height: 25)
-                        .background(Color.primary.opacity(0.06), in: RoundedRectangle(cornerRadius: 7, style: .continuous))
-                        .contentShape(RoundedRectangle(cornerRadius: 7, style: .continuous))
-                }
-                .buttonStyle(.plain)
-                .help("Back")
-                .accessibilityLabel("Back")
-                .pointerOnHover()
-                Text("Settings")
-                    .font(.system(size: 14, weight: .semibold))
-                Spacer()
-            }
-            SettingsPanel(store: store, updateChecker: updateChecker)
-        }
-        .padding(12)
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .top)
-    }
-}
-
-struct SettingsPanel: View {
-    @ObservedObject var store: UsageStore
-    @ObservedObject var updateChecker: UpdateChecker
-
-    var body: some View {
-        ScrollView {
-            VStack(alignment: .leading, spacing: 10) {
-                Toggle("Notifications", isOn: binding(\.notificationsEnabled))
-                    .toggleStyle(.switch)
-
-                Toggle("Do not disturb", isOn: Binding(
-                    get: { store.preferences.dndEnabled },
-                    set: { store.setDND($0) }
-                ))
-                .toggleStyle(.switch)
-
-                VStack(alignment: .leading, spacing: 6) {
-                    Text("Low quota threshold \(percentText(store.preferences.lowQuotaThreshold))")
-                        .font(.system(size: 11, weight: .semibold))
-                    Slider(value: binding(\.lowQuotaThreshold), in: 0.05...0.50, step: 0.05)
-                }
-
-                VStack(alignment: .leading, spacing: 6) {
-                    Text("Session warning \(percentText(store.preferences.sessionWarningThreshold))")
-                        .font(.system(size: 11, weight: .semibold))
-                    Slider(value: binding(\.sessionWarningThreshold), in: 0.05...0.40, step: 0.05)
-                }
-
-                Stepper("Cooldown \(store.preferences.notificationCooldownMinutes)m", value: intBinding(\.notificationCooldownMinutes), in: 5...360, step: 5)
-
-                Stepper("History \(store.preferences.historyRetentionDays)d", value: intBinding(\.historyRetentionDays), in: 1...60, step: 1)
-
-                Picker("Menu bar", selection: binding(\.menuBarMode)) {
-                    ForEach(MenuBarDisplayMode.allCases) { mode in
-                        Text(mode.label).tag(mode)
-                    }
-                }
-
-                if store.isAIInsightAvailable {
-                    Divider()
-                    AIInstructionsEditor(store: store)
-                }
-
-                Divider()
-
-                HStack(spacing: 8) {
-                    VStack(alignment: .leading, spacing: 2) {
-                        Text("App updates")
-                            .font(.system(size: 11, weight: .semibold))
-                        if updateChecker.isChecking {
-                            Text("Checking GitHub…")
-                                .foregroundStyle(.secondary)
-                        } else if let message = updateChecker.checkMessage {
-                            Text(message)
-                                .foregroundStyle(.secondary)
-                        } else if let date = updateChecker.lastCheckedAt {
-                            HStack(spacing: 3) {
-                                Text("Checked")
-                                Text(date, style: .relative)
-                            }
-                            .foregroundStyle(.secondary)
-                        } else {
-                            Text("Checks automatically every 6 hours")
-                                .foregroundStyle(.secondary)
-                        }
-                    }
-                    .font(.system(size: 10))
-
-                    Spacer()
-
-                    Button {
-                        updateChecker.checkNow()
-                    } label: {
-                        if updateChecker.isChecking {
-                            ProgressView()
-                                .controlSize(.small)
-                        } else {
-                            Text("Check now")
-                        }
-                    }
-                    .buttonStyle(.bordered)
-                    .controlSize(.small)
-                    .disabled(updateChecker.isChecking)
-                    .pointerOnHover()
-                }
-
-                Divider()
-
-                ConfigEditor(store: store)
-
-                HStack(spacing: 8) {
-                    Button {
-                        DiagnosticsReporter.presentSharePicker()
-                    } label: {
-                        Label("Send debug report", systemImage: "paperclip")
-                    }
-                    .buttonStyle(.borderedProminent)
-                    .controlSize(.small)
-                    .pointerOnHover()
-
-                    Button {
-                        DiagnosticsReporter.openLogFolder()
-                    } label: {
-                        Label("Logs", systemImage: "folder")
-                    }
-                    .buttonStyle(.bordered)
-                    .controlSize(.small)
-                    .pointerOnHover()
-                }
-            }
-            .font(.system(size: 12))
-            .padding(8)
-            .background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 8, style: .continuous))
-            .overlay(
-                RoundedRectangle(cornerRadius: 8, style: .continuous)
-                    .stroke(Color.primary.opacity(0.07), lineWidth: 1)
-            )
-        }
-        .frame(maxHeight: .infinity)
-    }
-
-    private func binding<Value>(_ keyPath: WritableKeyPath<AppPreferences, Value>) -> Binding<Value> {
-        Binding(
-            get: { store.preferences[keyPath: keyPath] },
-            set: { value in
-                var next = store.preferences
-                next[keyPath: keyPath] = value
-                store.updatePreferences(next)
-            }
-        )
-    }
-
-    private func intBinding(_ keyPath: WritableKeyPath<AppPreferences, Int>) -> Binding<Int> {
-        Binding(
-            get: { store.preferences[keyPath: keyPath] },
-            set: { value in
-                var next = store.preferences
-                next[keyPath: keyPath] = value
-                store.updatePreferences(next)
-            }
-        )
-    }
-}
-
-// In-app editor for the full config.json. Validates on save (rejecting invalid edits
-// before they reach disk) and reloads the store so changes take effect immediately.
-struct ConfigEditor: View {
-    @ObservedObject var store: UsageStore
-
-    @State private var text = ConfigLoader.rawContents()
-    @State private var error: String?
-    @State private var saved = false
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            HStack {
-                Text("Config JSON")
-                    .font(.system(size: 11, weight: .semibold))
-                Spacer()
-                Button {
-                    ConfigLoader.openInDefaultEditor()
-                } label: {
-                    Image(systemName: "arrow.up.forward.app")
-                }
-                .buttonStyle(.plain)
-                .help("Open in external editor")
-                .pointerOnHover()
-            }
-
-            TextEditor(text: $text)
-                .font(.system(size: 10, design: .monospaced))
-                .frame(height: 150)
-                .padding(4)
-                .background(Color.primary.opacity(0.04), in: RoundedRectangle(cornerRadius: 6, style: .continuous))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 6, style: .continuous)
-                        .stroke(Color.primary.opacity(0.1), lineWidth: 1)
-                )
-
-            if let error {
-                Text(error)
-                    .font(.system(size: 10))
-                    .foregroundStyle(.red)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-
-            HStack(spacing: 8) {
-                Button {
-                    save()
-                } label: {
-                    Label(saved ? "Saved" : "Save", systemImage: saved ? "checkmark" : "square.and.arrow.down")
-                }
-                .buttonStyle(.borderedProminent)
-                .controlSize(.small)
-                .pointerOnHover()
-
-                Button {
-                    text = ConfigLoader.rawContents()
-                    error = nil
-                    saved = false
-                } label: {
-                    Label("Revert", systemImage: "arrow.uturn.backward")
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-                .pointerOnHover()
-            }
-        }
-    }
-
-    private func save() {
-        do {
-            try ConfigLoader.saveRaw(text)
-            error = nil
-            saved = true
-            store.reloadConfig()
-            Task { @MainActor in
-                try? await Task.sleep(for: .seconds(1.5))
-                saved = false
-            }
-        } catch {
-            self.error = error.localizedDescription
-            saved = false
-        }
-    }
-}
-
-// Editor for the on-device AI prompt (config's `aiInstructions`). Empty means the built-in
-// default is used - shown as placeholder text. Saving persists and regenerates immediately.
-struct AIInstructionsEditor: View {
-    @ObservedObject var store: UsageStore
-
-    @State private var text: String
-    @State private var saved = false
-    @State private var error: String?
-
-    init(store: UsageStore) {
-        self.store = store
-        _text = State(initialValue: store.aiInstructions ?? "")
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Text("AI insight instructions")
-                .font(.system(size: 11, weight: .semibold))
-            Text("Steer how the on-device AI summarizes your usage. Leave empty for the default.")
-                .font(.system(size: 10))
-                .foregroundStyle(.secondary)
-                .fixedSize(horizontal: false, vertical: true)
-
-            ZStack(alignment: .topLeading) {
-                if text.isEmpty {
-                    Text(defaultAIInstructions)
-                        .font(.system(size: 10))
-                        .foregroundStyle(.tertiary)
-                        .padding(.horizontal, 8)
-                        .padding(.vertical, 8)
-                        .allowsHitTesting(false)
-                }
-                TextEditor(text: $text)
-                    .font(.system(size: 10))
-                    .frame(height: 90)
-                    .padding(4)
-                    .scrollContentBackground(.hidden)
-            }
-            .background(Color.primary.opacity(0.04), in: RoundedRectangle(cornerRadius: 6, style: .continuous))
-            .overlay(
-                RoundedRectangle(cornerRadius: 6, style: .continuous)
-                    .stroke(Color.primary.opacity(0.1), lineWidth: 1)
-            )
-
-            if let error {
-                Text(error)
-                    .font(.system(size: 10))
-                    .foregroundStyle(.red)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-
-            HStack(spacing: 8) {
-                Button {
-                    persist(text)
-                } label: {
-                    Label(saved ? "Saved" : "Save", systemImage: saved ? "checkmark" : "square.and.arrow.down")
-                }
-                .buttonStyle(.borderedProminent)
-                .controlSize(.small)
-                .pointerOnHover()
-
-                Button {
-                    text = ""
-                    persist("")
-                } label: {
-                    Label("Reset to default", systemImage: "arrow.uturn.backward")
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-                .disabled(text.isEmpty && store.aiInstructions == nil)
-                .pointerOnHover()
-            }
-        }
-    }
-
-    private func persist(_ value: String) {
-        if let failure = store.updateAIInstructions(value) {
-            error = failure
-            saved = false
-            return
-        }
-        error = nil
-        saved = true
-        Task { @MainActor in
-            try? await Task.sleep(for: .seconds(1.5))
-            saved = false
-        }
-    }
-}
-
-struct EmptyPanel: View {
-    var systemImage: String
-    var title: String
-    var detail: String
-
-    var body: some View {
-        VStack(spacing: 6) {
-            Image(systemName: systemImage)
-                .font(.system(size: 18, weight: .semibold))
-                .foregroundStyle(.secondary)
-            Text(title)
-                .font(.system(size: 12, weight: .semibold))
-            Text(detail)
-                .font(.system(size: 10))
-                .foregroundStyle(.secondary)
-                .multilineTextAlignment(.center)
-        }
-        .frame(maxWidth: .infinity)
-        .padding(18)
-        .background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 8, style: .continuous))
-        .overlay(
-            RoundedRectangle(cornerRadius: 8, style: .continuous)
-                .stroke(Color.primary.opacity(0.07), lineWidth: 1)
-        )
-    }
-}
